+++ test/Migrator.test.ts
@@ -13,7 +13,7 @@ describe("Migrator", function () {
     this.UniswapV2Pair = await ethers.getContractFactory("UniswapV2Pair")
     this.ERC20Mock = await ethers.getContractFactory("ERC20Mock", this.minter)
     this.FinaToken = await ethers.getContractFactory("FinaToken")
-    this.MasterChef = await ethers.getContractFactory("MasterChef")
+    this.FinaMaster = await ethers.getContractFactory("FinaMaster")
     this.Migrator = await ethers.getContractFactory("Migrator")
   })
 
@@ -41,7 +41,7 @@ describe("Migrator", function () {
 
     this.lp2 = await this.UniswapV2Pair.attach((await pair2.wait()).events[0].args.pair)
 
-    this.chef = await this.MasterChef.deploy(this.sushi.address, this.dev.address, "1000", "0", "100000")
+    this.chef = await this.FinaMaster.deploy(this.sushi.address, this.dev.address, "1000", "0", "100000")
     await this.chef.deployed()
 
     this.migrator = await this.Migrator.deploy(this.chef.address, this.factory1.address, this.factory2.address, "0")