+++ docs/HISTORY.md
@@ -3,8 +3,8 @@
 ## Early
 
 1. FinaToken Deployed
-2. MasterChef Deployed
-3. FinaToken Ownership Transfered to MasterChef
+2. FinaMaster Deployed
+3. FinaToken Ownership Transfered to FinaMaster
 4. TimeLock Deployed
 5. SushiFactory Deployed
 6. SushiRouter Deployed